Refactorizacion de las clases, sincronizacion en las tablas al momento de producir y consumir (GUI)
Se cambio la funcionalidad tanto de producir como consumir a la clase Buffer
Consumer solo se encarga de solicitar consumir
Producer solo se encarga de solicitar producir
Las sincronizacion al momento de producir un objeto y al momento de consumirlo ya son visibles
en la GUI
ProgressBar se encarga de representar el numero de objetos que hay en el buffer
JLabel9 se encarga de mostrar la cantidad de objetos consumidos.
